Atividade 01 adicionada 100%

+++ Linguagens1_Projetos/Atividade01/src/com/company/Contas.java
@@ -9,7 +9,7 @@ public class Contas {
         this.usuario = usuario;
         Contas.ID = Contas.ID + 1;
     }
-    public double sacar(double valor){
+    public int sacar(double valor){
         if(saldo >= valor){
             saldo -= valor;
             return 1;
